New components and some refactoring.
Added a theme to the material app. The theme for now only contains three colors. Designed and developed a ElevatedButtonComponent, to be used with all buttons. The idea is to achieve similar looks of all buttons, as well as to embrace reusability.
Designed and developed a drawer menu as well. The drawer menu required the development of a special top bar with a burger menu Icon as well, so this was done as well.
Finally, did some refactoring of the code as well. For example changed the signOutUser to be of type Either<AuthException, void> for better error handling