Onboarding, new re-usable component, login-ghost page
Proceeded further with the steps of the onboarding process. After account initialization (sign-up), the user is being navigated to a log-in page. The page is a ghost page, that can only be reached through the sign-up process. The page is assigned under the name LogInPageGhost (login_page_ghost.dart).
After sign-up, the first screen the user gets navigated to is the personal data page. The user is required to state their first and last name. This is the step, where the account is being initialised. In the database, the row that has the same id as the current user id, gets updated with the newly set values.

A new re-usable component was designed and developed too. The new component is a input field (TextField) with a custom heading above it. The component takes in three arguments, the heading, the placeholder text that shows inside the input field and the text editing controller to be assigned to the input field.