User_Profile class and Tunnel class update
Expanded the User_Profile class. Re-factored the default constructor of the class, the default constructor of the class takes no arguments. Re-defined the empty() constructor, changed from a constructor to a class method. The empty() method is used to initialize all of the class members with empty values. I as well added an onboarding method, which takes care of the first part of the onboarding process, gets the personal data of the user (names, email) and also creates a log of when the profile was created.
A clerProfile method was introduced as well. The clearProfile method sets all of the members to empty (the tunnelMACAddress, tunnelState and tunnelDevice as well). The purpose of the method is to clean the object members, once the user logs out.

Expanded the tunnel class as well. Added  separate methods for both assigning a device name to the object and assigning a mac address to the object.

Modified the uploadToDatabase() method inside the OnBoardingPersonalDataPage. Now, every time the method is called, a new table row is inserted with an owner_id, that is the same as the current authenticated user id.