UI, changed onboarding logic
Added a finishedOnBoarding member in the User_Profile class. The new member is supposed to track whether the onboarding process is complete. In further updates, this will change the logic, when the user opens the application.

Made some minior changes in the UI of the dashboard. Changed the font sizes and added back the drawer.