On boarding and supabase session
Write a code to init supabase within the main method of the application. This way the application again remebers the session and even if the user closes the application, when re-opened, the application brings them to home screen.

Modified the User_Profile class. Added a method, that fetches the data from the database and syncs it with the userProfile object in the client side.