worked with assigning tunnel id to profile
Worked with the process of assigning a ohmio device's id to the user profile in the database. When the user types in mac address, the application checks in the database if it exists and if it exists, gets the tunnel id and assigns its value to the profile > pintunnel_id. The code for this is stored inside the "retreive_tunnel_MAC_address.dart" file.

Added a splash screen to check if the user has finished the onboarding procees. The splash screen loads right after the user logs in.  The system retrieves the user profile, that has the same id as the current session user and checks if the user has the finishedOnBoarding set to true. If the user has the finshedOnBoarding in the database set to false, the user is being navigated to onboardingPersonalDataPage and needs to finish the onboarding. Else they get navigated to the dashboard page.