New UserData class, new method in supabaseManager and new component
Added a new method in the supabaseManager class, it's called getUserData(). The method gets all of the data about a user (names, username, email).

Wrote a userData class as well, part of the user_class.dart. It was specifically written for data binding with the getUserData method from the supabaseManager class.

Developed the drawer header widget as well. The widget just has a simple text that says hello to the user.
Changed the whole layout of the drawer menu in general, discovered different errors with it as well. Developed a component, that is to represent all of the different drawer items, but couldn't implement it, due to errors I couldn't fix.