Supabase channels and stream test + MQTT test
I wrote a stream and a channel, that both subscribe to the same supabase table and listen for changes. The demo for both is findable in the drawer menu.

I also wrote a MQTT test. The client conncets to the broker, subscribes to a specific table with authentication details and listens for changes.
In order to inrease code modularity and to separate the concerns, I moved the whole logic for the MQTT broker in a class called MqttManager.